Add files via upload
This project was originaly made for a science fair. We've made this project with the goal of recognizing drawings that a user makes in our site. We've done a database with images from the following things: apple, banana, cake, knife and axe. Afterwards we uploaded it to the TensorFlow that would recognise the drawings.